updated categories.js
