+++ src/api/categories.js
@@ -1,102 +1,123 @@
-'use strict';
-
-const categories = require('../categories');
-const events = require('../events');
-const user = require('../user');
-const groups = require('../groups');
-const privileges = require('../privileges');
-
-const categoriesAPI = module.exports;
-
-categoriesAPI.get = async function (caller, data) {
-    const [userPrivileges, category] = await Promise.all([
-        privileges.categories.get(data.cid, caller.uid),
-        categories.getCategoryData(data.cid),
-    ]);
-    if (!category || !userPrivileges.read) {
-        return null;
-    }
-
-    return category;
-};
-
-categoriesAPI.create = async function (caller, data) {
-    const response = await categories.create(data);
-    const categoryObjs = await categories.getCategories([response.cid], caller.uid);
-    return categoryObjs[0];
-};
-
-categoriesAPI.update = async function (caller, data) {
-    if (!data) {
-        throw new Error('[[error:invalid-data]]');
-    }
-    await categories.update(data);
-};
-
-categoriesAPI.delete = async function (caller, data) {
-    const name = await categories.getCategoryField(data.cid, 'name');
-    await categories.purge(data.cid, caller.uid);
-    await events.log({
-        type: 'category-purge',
-        uid: caller.uid,
-        ip: caller.ip,
-        cid: data.cid,
-        name: name,
+"use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-
-categoriesAPI.getPrivileges = async (caller, cid) => {
-    let responsePayload;
-
-    if (cid === 'admin') {
-        responsePayload = await privileges.admin.list(caller.uid);
-    } else if (!parseInt(cid, 10)) {
-        responsePayload = await privileges.global.list();
-    } else {
-        responsePayload = await privileges.categories.list(cid);
-    }
-
-    return responsePayload;
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
 };
-
-categoriesAPI.setPrivilege = async (caller, data) => {
-    const [userExists, groupExists] = await Promise.all([
-        user.exists(data.member),
-        groups.exists(data.member),
-    ]);
-
-    if (!userExists && !groupExists) {
-        throw new Error('[[error:no-user-or-group]]');
-    }
-    const privs = Array.isArray(data.privilege) ? data.privilege : [data.privilege];
-    const type = data.set ? 'give' : 'rescind';
-    if (!privs.length) {
-        throw new Error('[[error:invalid-data]]');
-    }
-    if (parseInt(data.cid, 10) === 0) {
-        const adminPrivList = await privileges.admin.getPrivilegeList();
-        const adminPrivs = privs.filter(priv => adminPrivList.includes(priv));
-        if (adminPrivs.length) {
-            await privileges.admin[type](adminPrivs, data.member);
+const categories_1 = __importDefault(require("../categories"));
+const events_1 = __importDefault(require("../events"));
+const user_1 = __importDefault(require("../user"));
+const groups_1 = __importDefault(require("../groups"));
+const privileges_1 = __importDefault(require("../privileges"));
+module.exports = function (categoriesAPI) {
+    categoriesAPI.get = function (caller, data) {
+        return __awaiter(this, void 0, void 0, function* () {
+            const privCat = yield Promise.all([
+                privileges_1.default.categories.get(data.cid, caller.uid),
+                // The next line calls a function in a module that has not been updated to TS yet
+                // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+                categories_1.default.getCategoryData(data.cid),
+            ]);
+            if (!privCat[1] || !privCat[0].read) {
+                return null;
+            }
+            return privCat[1];
+        });
+    };
+    categoriesAPI.create = function (caller, data) {
+        return __awaiter(this, void 0, void 0, function* () {
+            // The next line calls a function in a module that has not been updated to TS yet
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+            const response = yield categories_1.default.create(data);
+            const categoryObjs = yield categories_1.default.getCategories([response.cid], caller.uid);
+            return categoryObjs[0];
+        });
+    };
+    categoriesAPI.update = function (caller, data) {
+        return __awaiter(this, void 0, void 0, function* () {
+            if (!data) {
+                throw new Error('[[error:invalid-data]]');
+            }
+            // The next line calls a function in a module that has not been updated to TS yet
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+            yield categories_1.default.update(data);
+        });
+    };
+    categoriesAPI.delete = function (caller, data) {
+        return __awaiter(this, void 0, void 0, function* () {
+            // The next line calls a function in a module that has not been updated to TS yet
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+            const name = yield categories_1.default.getCategoryField(data.cid, 'name');
+            // The next line calls a function in a module that has not been updated to TS yet
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+            yield categories_1.default.purge(data.cid, caller.uid);
+            yield events_1.default.log({
+                type: 'category-purge',
+                uid: caller.uid,
+                ip: caller.ip,
+                cid: data.cid,
+                name: name,
+            });
+        });
+    };
+    categoriesAPI.getPrivileges = (caller, cid) => __awaiter(this, void 0, void 0, function* () {
+        let responsePayload;
+        if (cid === 'admin') {
+            responsePayload = yield privileges_1.default.admin.list(caller.uid);
         }
-        const globalPrivList = await privileges.global.getPrivilegeList();
-        const globalPrivs = privs.filter(priv => globalPrivList.includes(priv));
-        if (globalPrivs.length) {
-            await privileges.global[type](globalPrivs, data.member);
+        else if (!parseInt(cid, 10)) {
+            responsePayload = yield privileges_1.default.global.list();
         }
-    } else {
-        const categoryPrivList = await privileges.categories.getPrivilegeList();
-        const categoryPrivs = privs.filter(priv => categoryPrivList.includes(priv));
-        await privileges.categories[type](categoryPrivs, data.cid, data.member);
-    }
-
-    await events.log({
-        uid: caller.uid,
-        type: 'privilege-change',
-        ip: caller.ip,
-        privilege: data.privilege.toString(),
-        cid: data.cid,
-        action: data.set ? 'grant' : 'rescind',
-        target: data.member,
+        else {
+            responsePayload = yield privileges_1.default.categories.list(cid);
+        }
+        return responsePayload;
+    });
+    categoriesAPI.setPrivilege = (caller, data) => __awaiter(this, void 0, void 0, function* () {
+        const existsCheck = yield Promise.all([
+            user_1.default.exists(data.member),
+            groups_1.default.exists(data.member),
+        ]);
+        if (!existsCheck[0] && !existsCheck[1]) {
+            throw new Error('[[error:no-user-or-group]]');
+        }
+        const privs = Array.isArray(data.privilege) ? data.privilege : [data.privilege];
+        const type = data.set ? 'give' : 'rescind';
+        if (!privs.length) {
+            throw new Error('[[error:invalid-data]]');
+        }
+        if (parseInt(data.cid, 10) === 0) {
+            const adminPrivList = yield privileges_1.default.admin.getPrivilegeList();
+            const adminPrivs = privs.filter(priv => adminPrivList.includes(priv));
+            if (adminPrivs.length) {
+                yield privileges_1.default.admin[type](adminPrivs, data.member);
+            }
+            const globalPrivList = yield privileges_1.default.global.getPrivilegeList();
+            const globalPrivs = privs.filter(priv => globalPrivList.includes(priv));
+            if (globalPrivs.length) {
+                yield privileges_1.default.global[type](globalPrivs, data.member);
+            }
+        }
+        else {
+            const categoryPrivList = yield privileges_1.default.categories.getPrivilegeList();
+            const categoryPrivs = privs.filter(priv => categoryPrivList.includes(priv));
+            yield privileges_1.default.categories[type](categoryPrivs, data.cid, data.member);
+        }
+        yield events_1.default.log({
+            uid: caller.uid,
+            type: 'privilege-change',
+            ip: caller.ip,
+            privilege: data.privilege.toString(),
+            cid: data.cid,
+            action: data.set ? 'grant' : 'rescind',
+            target: data.member,
+        });
     });
 };
